+++ common/src/main/java/com/wisekrakr/firstgame/engine/gameobjects/enemies/enemyversions/EnemyFaceHugger.java
@@ -11,7 +11,7 @@ import java.util.Set;
 
 public class EnemyFaceHugger extends Enemy {
 
-    public EnemyFaceHugger(String name, Vector2 position, int health, float direction, float speed, float radius) {
+    public EnemyFaceHugger(String name, Vector2 position, double health, float direction, float speed, float radius) {
         super(GameObjectVisualizationType.FACE_HUGGER, name, position, health, direction, speed, radius);
 
         setAggroDistance(175f);
