+++ common/src/main/java/com/wisekrakr/firstgame/engine/gameobjects/enemies/enemyversions/EnemyShotty.java
@@ -11,7 +11,7 @@ import java.util.Set;
 
 public class EnemyShotty extends Enemy {
 
-    public EnemyShotty(String name, Vector2 position, int health, float direction, float speed, float radius) {
+    public EnemyShotty(String name, Vector2 position, double health, float direction, float speed, float radius) {
         super(GameObjectVisualizationType.SHOTTY, name, position, health, direction, speed, radius);
 
         setAggroDistance(212.5f);
